third_party/blink/web_tests/plugins/missing-plugin.html

<html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
</head>
<body>
<embed type="application/x-blink-test-plugin"></embed>
<embed type="application/x-non-existent"></embed>
<p>
    This tests that a missing plugin does not cause an assertion failure.
    <a href="http://webkit.org/b/41900">Bug 41900</a>
</p>
</body>
</html>
