third_party/blink/web_tests/platform/generic/plugins/missing-plugin-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
 

This tests that a missing plugin does not cause an assertion failure. Bug 41900
